refactor(tools): share async tool sync wrapper
